requirements.txt
mediapipe==0.10.14
opencv-python
#pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
#pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124
numpy==1.26.4
pandas==2.2.2
tensorflow==2.17.1
fastapi
uvicorn
python-dotenv
nltk
httpx
google.generativeai
colorama
#pip install "fastapi[standard]"